Unified signature for create_program_address
