+++ sdk/src/program.rs
@@ -2,38 +2,8 @@
 
 use crate::{
     account_info::AccountInfo, entrypoint::ProgramResult, entrypoint::SUCCESS,
-    instruction::Instruction, program_error::ProgramError, pubkey::Pubkey,
+    instruction::Instruction,
 };
-
-pub fn create_program_address(
-    seeds: &[&[u8]],
-    program_id: &Pubkey,
-) -> Result<Pubkey, ProgramError> {
-    let bytes = [
-        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-        0, 0,
-    ];
-    let result = unsafe {
-        sol_create_program_address(
-            seeds as *const _ as *const u8,
-            seeds.len() as u64,
-            program_id as *const _ as *const u8,
-            &bytes as *const _ as *const u8,
-        )
-    };
-    match result {
-        SUCCESS => Ok(Pubkey::new(&bytes)),
-        _ => Err(result.into()),
-    }
-}
-extern "C" {
-    fn sol_create_program_address(
-        seeds_addr: *const u8,
-        seeds_len: u64,
-        program_id_addr: *const u8,
-        address_bytes_addr: *const u8,
-    ) -> u64;
-}
 
 /// Invoke a cross-program instruction
 pub fn invoke(instruction: &Instruction, account_infos: &[AccountInfo]) -> ProgramResult {
